[BE-feat] 채팅기능, fword filter util 추가

+++ backend/board-service/src/main/java/com/a604/boardservice/repository/MessageRepository.java
@@ -11,5 +11,5 @@ public interface MessageRepository extends JpaRepository<Message, Long> {
      * @param messageListSeq
      * @return
      */
-    List<Message> findByMessageListSeqAndIsDeletedFalseOrderByCreatedAtAsc(int messageListSeq);
+    List<Message> findByMessageListSeqAndIsDeletedFalseOrderByCreatedAtAsc(long messageListSeq);
 }
